Move classes declared in ChangeFactory all to their own files
