+++ src/games/strategy/engine/data/changefactory/ChangeFactory.java
@@ -257,996 +257,3 @@ public class ChangeFactory {
 }
 
 
-/**
- * Resets the value to the default value.
- */
-class AttachmentPropertyReset extends Change {
-  private static final long serialVersionUID = 9208154387325299072L;
-  private final Attachable m_attachedTo;
-  private final String m_attachmentName;
-  private final Object m_oldValue;
-  private final String m_property;
-
-  AttachmentPropertyReset(final IAttachment attachment, final String property) {
-    if (attachment == null) {
-      throw new IllegalArgumentException("No attachment, property:" + property);
-    }
-    m_attachedTo = attachment.getAttachedTo();
-    m_attachmentName = attachment.getName();
-    m_oldValue = PropertyUtil.getPropertyFieldObject(property, attachment);
-    m_property = property;
-  }
-
-  AttachmentPropertyReset(final Attachable attachTo, final String attachmentName, final Object oldValue,
-      final String property) {
-    m_attachmentName = attachmentName;
-    m_attachedTo = attachTo;
-    m_oldValue = oldValue;
-    m_property = property;
-  }
-
-  public Attachable getAttachedTo() {
-    return m_attachedTo;
-  }
-
-  public String getAttachmentName() {
-    return m_attachmentName;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    final IAttachment attachment = m_attachedTo.getAttachment(m_attachmentName);
-    PropertyUtil.reset(m_property, attachment);
-  }
-
-  @Override
-  public Change invert() {
-    return new AttachmentPropertyResetUndo(m_attachedTo, m_attachmentName, m_oldValue, m_property);
-  }
-
-  @Override
-  public String toString() {
-    return "AttachmentPropertyClear attached to:" + m_attachedTo + " name:" + m_attachmentName + ", reset old value:"
-        + m_oldValue;
-  }
-}
-
-
-class AttachmentPropertyResetUndo extends Change {
-  private static final long serialVersionUID = 5943939650116851332L;
-  private final Attachable m_attachedTo;
-  private final String m_attachmentName;
-  private final Object m_newValue;
-  private final String m_property;
-
-  AttachmentPropertyResetUndo(final Attachable attachTo, final String attachmentName, final Object newValue,
-      final String property) {
-    m_attachmentName = attachmentName;
-    m_attachedTo = attachTo;
-    m_newValue = newValue;
-    m_property = property;
-  }
-
-  public Attachable getAttachedTo() {
-    return m_attachedTo;
-  }
-
-  public String getAttachmentName() {
-    return m_attachmentName;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    final IAttachment attachment = m_attachedTo.getAttachment(m_attachmentName);
-    PropertyUtil.set(m_property, m_newValue, attachment, false);
-  }
-
-  @Override
-  public Change invert() {
-    return new AttachmentPropertyReset(m_attachedTo, m_attachmentName, m_newValue, m_property);
-  }
-
-  @Override
-  public String toString() {
-    return "AttachmentPropertyClearUndo attached to:" + m_attachedTo + " name:" + m_attachmentName + " new value:"
-        + m_newValue;
-  }
-}
-
-
-/**
- * RelationshipChange this creates a change in relationshipType between two players, for example from Neutral to War.
- */
-class RelationshipChange extends Change {
-  private static final long serialVersionUID = 2694339584633196289L;
-  private final String m_player1;
-  private final String m_player2;
-  private final String m_OldRelation;
-  private final String m_NewRelation;
-
-  RelationshipChange(final PlayerID player1, final PlayerID player2, final RelationshipType oldRelation,
-      final RelationshipType newRelation) {
-    m_player1 = player1.getName();
-    m_player2 = player2.getName();
-    m_OldRelation = oldRelation.getName();
-    m_NewRelation = newRelation.getName();
-  }
-
-  private RelationshipChange(final String player1, final String player2, final String oldRelation,
-      final String newRelation) {
-    m_player1 = player1;
-    m_player2 = player2;
-    m_OldRelation = oldRelation;
-    m_NewRelation = newRelation;
-  }
-
-  @Override
-  public Change invert() {
-    return new RelationshipChange(m_player1, m_player2, m_NewRelation, m_OldRelation);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    data.getRelationshipTracker().setRelationship(data.getPlayerList().getPlayerID(m_player1),
-        data.getPlayerList().getPlayerID(m_player2), data.getRelationshipTypeList().getRelationshipType(m_NewRelation));
-    // now redraw territories in case of new hostility
-    if (Matches.RelationshipTypeIsAtWar.match(data.getRelationshipTypeList().getRelationshipType(m_NewRelation))) {
-      for (final Territory t : Match.getMatches(data.getMap().getTerritories(),
-          new CompositeMatchAnd<>(
-              Matches.territoryHasUnitsOwnedBy(data.getPlayerList().getPlayerID(m_player1)),
-              Matches.territoryHasUnitsOwnedBy(data.getPlayerList().getPlayerID(m_player2))))) {
-        t.notifyChanged();
-      }
-    }
-  }
-
-  @Override
-  public String toString() {
-    return "Add relation change. " + m_player1 + " and " + m_player2 + " change from " + m_OldRelation + " to "
-        + m_NewRelation;
-  }
-}
-
-
-/**
- * Add units
- */
-class AddUnits extends Change {
-  private static final long serialVersionUID = 2694342784633196289L;
-  private final String m_name;
-  private final Collection<Unit> m_units;
-  private final String m_type;
-
-  AddUnits(final UnitCollection collection, final Collection<Unit> units) {
-    m_units = new ArrayList<>(units);
-    m_name = collection.getHolder().getName();
-    m_type = collection.getHolder().getType();
-  }
-
-  AddUnits(final String name, final String type, final Collection<Unit> units) {
-    m_units = new ArrayList<>(units);
-    m_type = type;
-    m_name = name;
-  }
-
-  @Override
-  public Change invert() {
-    return new RemoveUnits(m_name, m_type, m_units);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final UnitHolder holder = data.getUnitHolder(m_name, m_type);
-    holder.getUnits().addAllUnits(m_units);
-  }
-
-  @Override
-  public String toString() {
-    return "Add unit change.  Add to:" + m_name + " units:" + m_units;
-  }
-}
-
-
-class RemoveUnits extends Change {
-  private static final long serialVersionUID = -6410444472951010568L;
-  private final String m_name;
-  private final Collection<Unit> m_units;
-  private final String m_type;
-
-  RemoveUnits(final UnitCollection collection, final Collection<Unit> units) {
-    this(collection.getHolder().getName(), collection.getHolder().getType(), units);
-  }
-
-  RemoveUnits(final String name, final String type, final Collection<Unit> units) {
-    m_units = new ArrayList<>(units);
-    m_name = name;
-    m_type = type;
-  }
-
-  @Override
-  public Change invert() {
-    return new AddUnits(m_name, m_type, m_units);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final UnitHolder holder = data.getUnitHolder(m_name, m_type);
-    if (!holder.getUnits().containsAll(m_units)) {
-      throw new IllegalStateException("Not all units present in:" + m_name + ".  Trying to remove:" + m_units
-          + " present:" + holder.getUnits().getUnits());
-    }
-    holder.getUnits().removeAllUnits(m_units);
-  }
-
-  @Override
-  public String toString() {
-    return "Remove unit change. Remove from:" + m_name + " units:" + m_units;
-  }
-}
-
-
-/**
- * Changes ownership of a territory.
- */
-class OwnerChange extends Change {
-  private static final long serialVersionUID = -5938125380623744929L;
-  /**
-   * Either new or old owner can be null.
-   */
-  private final String m_old;
-  private final String m_new;
-  private final String m_territory;
-
-  /**
-   * newOwner can be null
-   */
-  OwnerChange(final Territory territory, final PlayerID newOwner) {
-    m_territory = territory.getName();
-    m_new = getName(newOwner);
-    m_old = getName(territory.getOwner());
-  }
-
-  private OwnerChange(final String name, final String newOwner, final String oldOwner) {
-    m_territory = name;
-    m_new = newOwner;
-    m_old = oldOwner;
-  }
-
-  private String getName(final PlayerID player) {
-    if (player == null) {
-      return null;
-    }
-    return player.getName();
-  }
-
-  private PlayerID getPlayerID(final String name, final GameData data) {
-    if (name == null) {
-      return null;
-    }
-    return data.getPlayerList().getPlayerID(name);
-  }
-
-  @Override
-  public Change invert() {
-    return new OwnerChange(m_territory, m_old, m_new);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    // both names could be null
-    data.getMap().getTerritory(m_territory).setOwner(getPlayerID(m_new, data));
-  }
-
-  @Override
-  public String toString() {
-    return m_new + " takes " + m_territory + " from " + m_old;
-  }
-}
-
-
-/**
- * Changes ownership of a unit.
- */
-class PlayerOwnerChange extends Change {
-  /**
-   * Maps unit id -> owner as String
-   */
-  private final Map<GUID, String> m_old;
-  private final Map<GUID, String> m_new;
-  private final String m_location;
-  private static final long serialVersionUID = -9154938431233632882L;
-
-  PlayerOwnerChange(final Collection<Unit> units, final PlayerID newOwner, final Territory location) {
-    m_old = new HashMap<>();
-    m_new = new HashMap<>();
-    m_location = location.getName();
-    for (final Unit unit : units) {
-      m_old.put(unit.getID(), unit.getOwner().getName());
-      m_new.put(unit.getID(), newOwner.getName());
-    }
-  }
-
-  PlayerOwnerChange(final Map<GUID, String> newOwner, final Map<GUID, String> oldOwner, final String location) {
-    m_old = oldOwner;
-    m_new = newOwner;
-    m_location = location;
-  }
-
-  @Override
-  public Change invert() {
-    return new PlayerOwnerChange(m_old, m_new, m_location);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    for (final GUID id : m_new.keySet()) {
-      final Unit unit = data.getUnits().get(id);
-      if (!m_old.get(id).equals(unit.getOwner().getName())) {
-        throw new IllegalStateException("Wrong owner, expecting" + m_old.get(id) + " but got " + unit.getOwner());
-      }
-      final String owner = m_new.get(id);
-      final PlayerID player = data.getPlayerList().getPlayerID(owner);
-      unit.setOwner(player);
-    }
-    data.getMap().getTerritory(m_location).notifyChanged();
-  }
-
-  @Override
-  public String toString() {
-    return "Some units change owners in territory " + m_location;
-  }
-}
-
-/**
- * Adds/removes resource from a player.
- */
-class ChangeResourceChange extends Change {
-  private static final long serialVersionUID = -2304294240555842126L;
-  private final String m_player;
-  private final String m_resource;
-  private final int m_quantity;
-
-  ChangeResourceChange(final PlayerID player, final Resource resource, final int quantity) {
-    m_player = player.getName();
-    m_resource = resource.getName();
-    m_quantity = quantity;
-  }
-
-  private ChangeResourceChange(final String player, final String resource, final int quantity) {
-    m_player = player;
-    m_resource = resource;
-    m_quantity = quantity;
-  }
-
-  @Override
-  public Change invert() {
-    return new ChangeResourceChange(m_player, m_resource, -m_quantity);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final Resource resource = data.getResourceList().getResource(m_resource);
-    final ResourceCollection resources = data.getPlayerList().getPlayerID(m_player).getResources();
-    if (m_quantity > 0) {
-      resources.addResource(resource, m_quantity);
-    } else if (m_quantity < 0) {
-      resources.removeResource(resource, -m_quantity);
-    }
-  }
-
-  @Override
-  public String toString() {
-    return "Change resource.  Resource:" + m_resource + " quantity:" + m_quantity + " Player:" + m_player;
-  }
-}
-
-
-class SetPropertyChange extends Change {
-  private static final long serialVersionUID = -1377597975513821508L;
-  private final String m_property;
-  private final Object m_value;
-  private final Object m_oldValue;
-
-  SetPropertyChange(final String property, final Object value, final GameProperties properties) {
-    m_property = property;
-    m_value = value;
-    m_oldValue = properties.get(property);
-  }
-
-  private SetPropertyChange(final String property, final Object value, final Object oldValue) {
-    m_property = property;
-    m_value = value;
-    m_oldValue = oldValue;
-  }
-
-  @Override
-  public Change invert() {
-    return new SetPropertyChange(m_property, m_oldValue, m_value);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    data.getProperties().set(m_property, m_value);
-  }
-}
-
-
-class AddProductionRule extends Change {
-  private static final long serialVersionUID = 2583955907289570063L;
-  private final ProductionRule m_rule;
-  private final ProductionFrontier m_frontier;
-
-  public AddProductionRule(final ProductionRule rule, final ProductionFrontier frontier) {
-    if (rule == null) {
-      throw new IllegalArgumentException("Null rule");
-    }
-    if (frontier == null) {
-      throw new IllegalArgumentException("Null frontier");
-    }
-    m_rule = rule;
-    m_frontier = frontier;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    m_frontier.addRule(m_rule);
-  }
-
-  @Override
-  public Change invert() {
-    return new RemoveProductionRule(m_rule, m_frontier);
-  }
-}
-
-
-class RemoveProductionRule extends Change {
-  private static final long serialVersionUID = 2312599802275503095L;
-  private final ProductionRule m_rule;
-  private final ProductionFrontier m_frontier;
-
-  public RemoveProductionRule(final ProductionRule rule, final ProductionFrontier frontier) {
-    if (rule == null) {
-      throw new IllegalArgumentException("Null rule");
-    }
-    if (frontier == null) {
-      throw new IllegalArgumentException("Null frontier");
-    }
-    m_rule = rule;
-    m_frontier = frontier;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    m_frontier.removeRule(m_rule);
-  }
-
-  @Override
-  public Change invert() {
-    return new AddProductionRule(m_rule, m_frontier);
-  }
-}
-
-
-class AddAvailableTech extends Change {
-  private static final long serialVersionUID = 5664428883866434959L;
-  private final TechAdvance m_tech;
-  private final TechnologyFrontier m_frontier;
-  private final PlayerID m_player;
-
-  public AddAvailableTech(final TechnologyFrontier front, final TechAdvance tech, final PlayerID player) {
-    if (front == null) {
-      throw new IllegalArgumentException("Null tech category");
-    }
-    if (tech == null) {
-      throw new IllegalArgumentException("Null tech");
-    }
-    m_tech = tech;
-    m_frontier = front;
-    m_player = player;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    final TechnologyFrontier front = m_player.getTechnologyFrontierList().getTechnologyFrontier(m_frontier.getName());
-    front.addAdvance(m_tech);
-  }
-
-  @Override
-  public Change invert() {
-    return new RemoveAvailableTech(m_frontier, m_tech, m_player);
-  }
-}
-
-
-class RemoveAvailableTech extends Change {
-  private static final long serialVersionUID = 6131447662760022521L;
-  private final TechAdvance m_tech;
-  private final TechnologyFrontier m_frontier;
-  private final PlayerID m_player;
-
-  public RemoveAvailableTech(final TechnologyFrontier front, final TechAdvance tech, final PlayerID player) {
-    if (front == null) {
-      throw new IllegalArgumentException("Null tech category");
-    }
-    if (tech == null) {
-      throw new IllegalArgumentException("Null tech");
-    }
-    m_tech = tech;
-    m_frontier = front;
-    m_player = player;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    final TechnologyFrontier front = m_player.getTechnologyFrontierList().getTechnologyFrontier(m_frontier.getName());
-    front.removeAdvance(m_tech);
-  }
-
-  @Override
-  public Change invert() {
-    return new AddAvailableTech(m_frontier, m_tech, m_player);
-  }
-}
-
-
-class AddAttachmentChange extends Change {
-  private static final long serialVersionUID = -21015135248288454L;
-  private final IAttachment m_attachment;
-  private final String m_originalAttachmentName;
-  private final Attachable m_originalAttachable;
-  private final Attachable m_attachable;
-  private final String m_name;
-
-  public AddAttachmentChange(final IAttachment attachment, final Attachable attachable, final String name) {
-    m_attachment = attachment;
-    m_originalAttachmentName = attachment.getName();
-    m_originalAttachable = attachment.getAttachedTo();
-    m_attachable = attachable;
-    m_name = name;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    m_attachable.addAttachment(m_name, m_attachment);
-    m_attachment.setName(m_name);
-    m_attachment.setAttachedTo(m_attachable);
-  }
-
-  @Override
-  public Change invert() {
-    return new RemoveAttachmentChange(m_attachment, m_originalAttachable, m_originalAttachmentName);
-  }
-}
-
-
-class RemoveAttachmentChange extends Change {
-  private static final long serialVersionUID = 6365648682759047674L;
-  private final IAttachment m_attachment;
-  private final String m_originalAttachmentName;
-  private final Attachable m_originalAttachable;
-  private final Attachable m_attachable;
-  private final String m_name;
-
-  public RemoveAttachmentChange(final IAttachment attachment, final Attachable attachable, final String name) {
-    m_attachment = attachment;
-    m_originalAttachmentName = attachment.getName();
-    m_originalAttachable = attachment.getAttachedTo();
-    m_attachable = attachable;
-    m_name = name;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    m_originalAttachable.removeAttachment(m_originalAttachmentName);
-    m_attachment.setAttachedTo(m_attachable);
-    m_attachment.setName(m_name);
-    if (m_attachable != null && m_name != null) {
-      m_attachable.addAttachment(m_name, m_attachment);
-    }
-  }
-
-  @Override
-  public Change invert() {
-    return new AddAttachmentChange(m_attachment, m_originalAttachable, m_originalAttachmentName);
-  }
-}
-
-
-class AddGridGameMapChange extends Change {
-  private static final long serialVersionUID = -8326219690181895908L;
-  final Map<Territory, Set<Territory>> m_removeTerritoriesAndConnections;
-  final Map<Territory, Set<Territory>> m_addTerritoriesAndConnections;
-  final String m_gridType;
-  final String m_name;
-  final int m_xs;
-  final int m_ys;
-  final Set<String> m_water;
-  final Set<String> m_oldWater;
-  final String m_horizontalConnections;
-  final String m_verticalConnections;
-  final String m_diagonalConnections;
-  final int m_oldXs;
-  final int m_oldYs;
-
-  public AddGridGameMapChange(final GameMap map, final String gridType, final String name, final int xs, final int ys,
-      final Set<String> water, final String horizontalConnections, final String verticalConnections,
-      final String diagonalConnections) {
-    m_oldXs = map.getXDimension();
-    m_oldYs = map.getYDimension();
-    m_gridType = gridType;
-    m_name = name;
-    m_xs = xs;
-    m_ys = ys;
-    m_water = water;
-    m_horizontalConnections = horizontalConnections;
-    m_verticalConnections = verticalConnections;
-    m_diagonalConnections = diagonalConnections;
-    m_oldWater = new HashSet<>();
-    for (final Territory t : map.getTerritories()) {
-      if (t.isWater()) {
-        m_oldWater.add(t.getName());
-      }
-    }
-    m_removeTerritoriesAndConnections = new HashMap<>();
-    m_addTerritoriesAndConnections = new HashMap<>();
-    if (map.getXDimension() > m_xs) {
-      // we will be removing territories
-      for (final Territory t : map.getTerritories()) {
-        String tname = t.getName().replaceFirst(m_name + "_", "");
-        tname = tname.substring(0, tname.indexOf("_"));
-        final int tx = Integer.parseInt(tname);
-        if (tx > m_xs - 1) {
-          m_removeTerritoriesAndConnections.put(t, map.getNeighbors(t));
-        }
-      }
-    } else if (map.getXDimension() < m_xs) {
-      // adding territories
-    }
-    if (map.getYDimension() > m_ys) {
-      // we will be removing territories
-      for (final Territory t : map.getTerritories()) {
-        String tname = t.getName().replaceFirst(m_name + "_", "");
-        tname = tname.substring(tname.indexOf("_") + 1, tname.length());
-        final int ty = Integer.parseInt(tname);
-        if (ty > m_ys - 1) {
-          m_removeTerritoriesAndConnections.put(t, map.getNeighbors(t));
-        }
-      }
-    } else if (map.getYDimension() < m_ys) {
-      // adding territories
-    }
-  }
-
-  public AddGridGameMapChange(final String gridType, final String name, final int oldXs, final int oldYs,
-      final int newXs, final int newYs, final Set<String> oldWater, final Set<String> water,
-      final String horizontalConnections, final String verticalConnections, final String diagonalConnections,
-      final Map<Territory, Set<Territory>> removeTerritoriesAndConnections,
-      final Map<Territory, Set<Territory>> addTerritoriesAndConnections) {
-    m_oldXs = oldXs;
-    m_oldYs = oldYs;
-    m_gridType = gridType;
-    m_name = name;
-    m_xs = newXs;
-    m_ys = newYs;
-    m_oldWater = oldWater;
-    m_water = water;
-    m_horizontalConnections = horizontalConnections;
-    m_verticalConnections = verticalConnections;
-    m_diagonalConnections = diagonalConnections;
-    m_removeTerritoriesAndConnections = removeTerritoriesAndConnections;
-    m_addTerritoriesAndConnections = addTerritoriesAndConnections;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    if (data.getMap().getXDimension() != m_xs || data.getMap().getYDimension() != m_ys) {
-      try {
-        final GameMap map = data.getMap();
-        for (final Territory t : m_removeTerritoriesAndConnections.keySet()) {
-          map.removeTerritory(t);
-        }
-        GameParser.setGrids(data, m_gridType, m_name, Integer.toString(m_xs), Integer.toString(m_ys), m_water,
-            m_horizontalConnections, m_verticalConnections, m_diagonalConnections, true);
-        map.notifyChanged();
-      } catch (final Exception e) {
-        ClientLogger.logQuietly(e);
-      }
-    }
-  }
-
-  @Override
-  public Change invert() {
-    return new AddGridGameMapChange(m_gridType, m_name, m_xs, m_ys, m_oldXs, m_oldYs, m_water, m_oldWater,
-        m_horizontalConnections, m_verticalConnections, m_diagonalConnections, m_addTerritoriesAndConnections,
-        m_removeTerritoriesAndConnections);
-  }
-}
-
-
-/**
- * Change a players production frontier.
- */
-class ProductionFrontierChange extends Change {
-  private final String m_startFrontier;
-  private final String m_endFrontier;
-  private final String m_player;
-  private static final long serialVersionUID = 3336145814067456701L;
-
-  ProductionFrontierChange(final ProductionFrontier newFrontier, final PlayerID player) {
-    m_startFrontier = player.getProductionFrontier().getName();
-    m_endFrontier = newFrontier.getName();
-    m_player = player.getName();
-  }
-
-  ProductionFrontierChange(final String startFrontier, final String endFrontier, final String player) {
-    m_startFrontier = startFrontier;
-    m_endFrontier = endFrontier;
-    m_player = player;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final PlayerID player = data.getPlayerList().getPlayerID(m_player);
-    final ProductionFrontier frontier = data.getProductionFrontierList().getProductionFrontier(m_endFrontier);
-    player.setProductionFrontier(frontier);
-  }
-
-  @Override
-  public Change invert() {
-    return new ProductionFrontierChange(m_endFrontier, m_startFrontier, m_player);
-  }
-}
-
-
-class GameSequenceChange extends Change {
-  private static final long serialVersionUID = -8925565771506676074L;
-  private final GameStep[] m_oldSteps;
-  private final GameStep[] m_newSteps;
-
-  GameSequenceChange(final GameSequence oldSequence, final GameStep[] newSteps) {
-    final ArrayList<GameStep> oldSteps = new ArrayList<>();
-    for (final GameStep step : oldSequence) {
-      oldSteps.add(step);
-    }
-    m_oldSteps = oldSteps.toArray(new GameStep[oldSteps.size()]);
-    m_newSteps = newSteps;
-  }
-
-  private GameSequenceChange(final GameStep[] oldSteps, final GameStep[] newSteps) {
-    m_oldSteps = oldSteps;
-    m_newSteps = newSteps;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final GameSequence steps = data.getSequence();
-    steps.removeAllSteps();
-    for (final GameStep newStep : m_newSteps) {
-      steps.addStep(newStep);
-    }
-  }
-
-  @Override
-  public Change invert() {
-    return new GameSequenceChange(m_newSteps, m_oldSteps);
-  }
-}
-
-
-class ObjectPropertyChange extends Change {
-  private static final long serialVersionUID = 4218093376094170940L;
-  private final Object m_object;
-  private String m_property;
-  private Object m_newValue;
-  private Object m_oldValue;
-
-  public ObjectPropertyChange(final Object object, final String property, final Object newValue) {
-    m_object = object;
-    m_property = property.intern();
-    m_newValue = newValue;
-    // m_oldValue = PropertyUtil.get(property, object);
-    m_oldValue = PropertyUtil.getPropertyFieldObject(property, object);
-  }
-
-  public ObjectPropertyChange(final Object object, final String property, final Object newValue,
-      final Object oldValue) {
-    m_object = object;
-    // prevent multiple copies of the property names being held in the game
-    m_property = property.intern();
-    m_newValue = newValue;
-    m_oldValue = oldValue;
-  }
-
-  private void readObject(final ObjectInputStream stream) throws IOException, ClassNotFoundException {
-    stream.defaultReadObject();
-    m_property = m_property.intern();
-  }
-
-  @Override
-  public Change invert() {
-    return new ObjectPropertyChange(m_object, m_property, m_oldValue, m_newValue);
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    PropertyUtil.set(m_property, m_newValue, m_object);
-  }
-
-  @Override
-  public String toString() {
-    return "Property change, unit:" + m_object + " property:" + m_property + " newValue:" + m_newValue + " oldValue:"
-        + m_oldValue;
-  }
-}
-
-
-class GenericTechChange extends Change {
-  private static final long serialVersionUID = -2439447526511535571L;
-  private final Attachable m_attachedTo;
-  private final String m_attachmentName;
-  private final Boolean m_newValue;
-  private final Boolean m_oldValue;
-  private final String m_property;
-
-  public Attachable getAttachedTo() {
-    return m_attachedTo;
-  }
-
-  public String getAttachmentName() {
-    return m_attachmentName;
-  }
-
-  GenericTechChange(final TechAttachment attachment, final Boolean newValue, final String property) {
-    if (attachment == null) {
-      throw new IllegalArgumentException("No attachment, newValue:" + newValue + " property:" + property);
-    }
-    m_attachedTo = attachment.getAttachedTo();
-    m_attachmentName = attachment.getName();
-    m_oldValue = Boolean.valueOf(attachment.hasGenericTech(property));
-    m_newValue = Boolean.valueOf(newValue);
-    m_property = property;
-  }
-
-  public GenericTechChange(final Attachable attachTo, final String attachmentName, final Boolean newValue,
-      final Boolean oldValue, final String property) {
-    m_attachmentName = attachmentName;
-    m_attachedTo = attachTo;
-    m_newValue = newValue;
-    m_oldValue = oldValue;
-    m_property = property;
-  }
-
-  @Override
-  public void perform(final GameData data) {
-    final TechAttachment attachment = (TechAttachment) m_attachedTo.getAttachment(m_attachmentName);
-    attachment.setGenericTech(m_property, m_newValue);
-  }
-
-  @Override
-  public Change invert() {
-    return new GenericTechChange(m_attachedTo, m_attachmentName, m_oldValue, m_newValue, m_property);
-  }
-
-  @Override
-  public String toString() {
-    return "GenericTechChange attached to:" + m_attachedTo + " name:" + m_attachmentName + " new value:" + m_newValue
-        + " old value:" + m_oldValue;
-  }
-}
-
-
-class AddBattleRecordsChange extends Change {
-  private static final long serialVersionUID = -6927678548172402611L;
-  private final BattleRecords m_recordsToAdd;
-  private final int m_round;
-
-  AddBattleRecordsChange(final BattleRecords battleRecords, final GameData data) {
-    m_round = data.getSequence().getRound();
-    // make a copy because this is only done once, and only externally from battle
-    // tracker, and the source will be cleared (battle tracker clears out the records
-    // each turn)
-    m_recordsToAdd = new BattleRecords(battleRecords);
-  }
-
-  AddBattleRecordsChange(final BattleRecords battleRecords, final int round) {
-    m_round = round;
-    // do not make a copy, this is only called from RemoveBattleRecordsChange, and we make a copy when we
-    // perform, so no need for another copy.
-    m_recordsToAdd = battleRecords;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final Map<Integer, BattleRecords> currentRecords = data.getBattleRecordsList().getBattleRecordsMap();
-    // make a copy because otherwise ours will be
-    // cleared when we RemoveBattleRecordsChange
-    BattleRecordsList.addRecords(currentRecords, m_round, new BattleRecords(m_recordsToAdd));
-  }
-
-  @Override
-  public Change invert() {
-    return new RemoveBattleRecordsChange(m_recordsToAdd, m_round);
-  }
-
-  @Override
-  public String toString() {
-    // This only occurs when serialization went badly, or something cannot be serialized.
-    if (m_recordsToAdd == null) {
-      throw new IllegalStateException(
-          "Records cannot be null (most likely caused by improper or impossible serialization): " + m_recordsToAdd);
-    }
-    return "Adding Battle Records: " + m_recordsToAdd;
-  }
-}
-
-
-class RemoveBattleRecordsChange extends Change {
-  private static final long serialVersionUID = 3286634991233029854L;
-  private final BattleRecords m_recordsToRemove;
-  private final int m_round;
-
-  RemoveBattleRecordsChange(final BattleRecords battleRecords, final int round) {
-    m_round = round;
-    // do not make a copy, this is only called from AddBattleRecordsChange, and we make a copy when we
-    // perform, so no need for another copy.
-    m_recordsToRemove = battleRecords;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final Map<Integer, BattleRecords> currentRecords = data.getBattleRecordsList().getBattleRecordsMap();
-    // make a copy else we will get a concurrent modification error
-    BattleRecordsList.removeRecords(currentRecords, m_round, new BattleRecords(m_recordsToRemove));
-  }
-
-  @Override
-  public Change invert() {
-    return new AddBattleRecordsChange(m_recordsToRemove, m_round);
-  }
-
-  @Override
-  public String toString() {
-    // This only occurs when serialization went badly, or something cannot be serialized.
-    if (m_recordsToRemove == null) {
-      throw new IllegalStateException(
-          "Records cannot be null (most likely caused by improper or impossible serialization): " + m_recordsToRemove);
-    }
-    return "Adding Battle Records: " + m_recordsToRemove;
-  }
-}
-
-
-class PlayerWhoAmIChange extends Change {
-  private static final long serialVersionUID = -1486914230174337300L;
-  private final String m_startWhoAmI;
-  private final String m_endWhoAmI;
-  private final String m_player;
-
-  PlayerWhoAmIChange(final String newWhoAmI, final PlayerID player) {
-    m_startWhoAmI = player.getWhoAmI();
-    m_endWhoAmI = newWhoAmI;
-    m_player = player.getName();
-  }
-
-  PlayerWhoAmIChange(final String startWhoAmI, final String endWhoAmI, final String player) {
-    m_startWhoAmI = startWhoAmI;
-    m_endWhoAmI = endWhoAmI;
-    m_player = player;
-  }
-
-  @Override
-  protected void perform(final GameData data) {
-    final PlayerID player = data.getPlayerList().getPlayerID(m_player);
-    player.setWhoAmI(m_endWhoAmI);
-  }
-
-  @Override
-  public Change invert() {
-    return new PlayerWhoAmIChange(m_endWhoAmI, m_startWhoAmI, m_player);
-  }
-
-  @Override
-  public String toString() {
-    return m_player + " changed from " + m_startWhoAmI + " to " + m_endWhoAmI;
-  }
-}
